+++ jest.config.js
@@ -21,7 +21,7 @@ module.exports = {
   // transform: {
   //   ".*\\.js$": "babel-jest",
   // },
-  transformIgnorePatterns: ['<rootDir>/node_modules/'],
+  transformIgnorePatterns: ['<rootDir>/node_modules/(?!d3|d3-array|internmap|delaunator|robust-predicates)'],
   testResultsProcessor: 'jest-sonar-reporter',
   collectCoverage: true,
   collectCoverageFrom: ['src/**/*.js'],