+++ changelog.md
@@ -16,8 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve all class models and their individual tests.
 - Remove function renderString for SVG template.
 - Create default class structure draw/metadata/parser/render and their individual tests.
-- Add default error for parser
-- Add `Draw` and `DrawDefaultContent` methods to DefaultDrawer.js
+- Add default error for parser.
+- Add `draw` and `drawDefaultModel` methods to DefaultDrawer.js.
+- Add functions to set default component position.
 
 ## [0.1.0] - 2022-07-07