+++ changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-###Added
+### Added
 
 - Workflow to deploy documentation on Github pages
 
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `draw` and `drawDefaultModel` methods to DefaultDrawer.js.
 - Add functions to set default component position.
 - Add `moveComponent` method to move svg components in modelizer.
+- Have specific draw method for template
 
 ## [0.1.0] - 2022-07-07